Actualiza estilos y servicios

+++ src/app/pages/admin/departament/departament-card/departament-card.component.ts
@@ -1,14 +1,17 @@
 import { Component, EventEmitter, Inject, Input, OnInit, Output, inject } from '@angular/core';
 import { IDepartment } from '../../../../core/models/types';
 import { DepartamentEditComponent } from '../departament-edit/departament-edit.component';
-import { DepartamentService } from '../../../../core/services/departament.service';
-import { MessageService } from 'primeng/api';
+import { ConfirmationService, MessageService } from 'primeng/api';
+import { ButtonModule } from 'primeng/button';
+import { ConfirmDialog, ConfirmDialogModule } from 'primeng/confirmdialog';
+import { DialogService, DynamicDialogRef } from 'primeng/dynamicdialog';
+import { DepartamentService } from '../../../../core/services/admin/departament.service';
 
 
 @Component({
   selector: 'app-departament-card',
   standalone: true,
-  imports: [DepartamentEditComponent],
+  imports: [ButtonModule,],
   templateUrl: './departament-card.component.html',
   styleUrl: './departament-card.component.css'
 })
@@ -21,10 +24,16 @@ export class DepartamentCardComponent {
 
    }
 
+   
+  ref: DynamicDialogRef | undefined;
+
+
    @Output() departamentDelete: EventEmitter<IDepartment> = new EventEmitter<IDepartment>();
 
    departamentService = inject(DepartamentService) 
-   messageServive = inject(MessageService)
+   confirmationService = inject(ConfirmationService)
+   dialogService = inject(DialogService)
+   messageService = inject(MessageService)
 
 
 
@@ -32,26 +41,68 @@ export class DepartamentCardComponent {
     this.departament = departament
    }
 
+   
 
 
-   handleDelete(){
-   if(this.departament.id_departamento){
-    this.departamentService.deleteDepartament(this.departament.id_departamento).subscribe((response) => {
-      this.messageServive.add({
-        severity:'success',
-        summary:'Departamento eliminado',
-        detail:'Departamento eliminado correctamente'
-      })
-      this.departamentService.updateDepartamentCache()
-      this.departamentDelete.emit(this.departament)
-    },
-    (error) => {
-      this.messageServive.add({severity:'error', summary:'Error', detail:'Error al eliminar Departamento'})
-      console.error('Error editing License:', error);
-    }
-    )
+
+   editDepartament(){
+    this.ref = this.dialogService.open(DepartamentEditComponent, {
+      header: 'Editar Departamento',
+      width: 'auto',
+      height: 'auto',
+      contentStyle: { overflow: 'auto', 'padding':'0'},
+      breakpoints: {
+        '960px': '75vw',
+        '640px': '90vw',
+      },
+      data: {
+        departament: this.departament,
+      }
+    });
+
+    this.ref.onClose.subscribe((data: any) => {});
+
+    this.ref.onMaximize.subscribe((value) => {
+      this.messageService.add({
+        severity: 'info',
+        summary: 'Maximized',
+        detail: `maximized: ${value.maximized}`,
+      });
+    });
    }
 
-    
+   confirmDelete(event: Event){
+    this.confirmationService.confirm({
+      target: event.target as EventTarget,
+      message: 'Estas seguro de eliminar este departamento?, si lo haces todos los empleados asociados a el desapareceran',
+      header: 'Confirmar Eliminacion',
+      icon: 'pi pi-info-circle',
+      acceptButtonStyleClass:"p-button-danger p-button-text",
+      rejectButtonStyleClass:"p-button-text p-button-text",
+      acceptIcon:"none",
+      rejectIcon:"none",
+
+
+      accept: () => {
+        if(this.departament.id_departamento){
+          this.departamentService.deleteDepartament(this.departament.id_departamento).subscribe((response) => {
+            this.messageService.add({
+              severity:'success',
+              summary:'Departamento eliminado',
+              detail:'Departamento eliminado correctamente'
+            })
+          },
+          (error) => {
+            this.messageService.add({severity:'error', summary:'Error', detail:'Error al eliminar Departamento'})
+            console.error('Error editing License:', error);
+          }
+          )
+         }
+      
+      },
+      reject: () => {
+          this.messageService.add({ severity: 'error', summary: 'Cancelado', detail: 'Haz cancelado la eliminacion' });
+      }
+  });
    }
 }
